+++ example/TeleDart_example.dart
@@ -46,6 +46,11 @@ void main() {
           'https://raw.githubusercontent.com/DinoLeung/TeleDart/master/example/dash_paper_plane.png',
           caption: 'This is how Dash found the paper plane'));
 
+  // Sick of boilerplates? Reply messages like below, nice and tidy
+  teledart
+      .onMessage(keyword: 'Fight for freedom')
+      .listen((message) => message.replyMessage('Stand with Hong Kong'));
+
   // Inline mode.
   teledart
       .onInlineQuery()